+++ src/test/groovy/com/citytechinc/cq/groovyconsole/servlets/ScriptPostServletSpec.groovy
@@ -2,6 +2,8 @@ package com.citytechinc.cq.groovyconsole.servlets
 import com.citytechinc.cq.groovy.extension.services.OsgiComponentService
 import com.citytechinc.cq.groovyconsole.services.ConfigurationService
 import com.citytechinc.cq.groovyconsole.services.EmailService
+import com.citytechinc.cq.groovyconsole.services.GroovyRunService
+import com.citytechinc.cq.groovyconsole.services.impl.GroovyRunServiceImpl
 import com.day.cq.replication.Replicator
 import com.day.cq.search.QueryBuilder
 import groovy.json.JsonSlurper
@@ -14,17 +16,23 @@ class ScriptPostServletSpec extends AbstractServletSpec {
 
     @Shared servlet
 
+    @Shared service
+
     @Shared script
 
     def setupSpec() {
         servlet = new ScriptPostServlet()
 
-        servlet.replicator = Mock(Replicator)
-        servlet.componentService = Mock(OsgiComponentService)
-        servlet.bundleContext = Mock(BundleContext)
-        servlet.configurationService = Mock(ConfigurationService)
-        servlet.queryBuilder = Mock(QueryBuilder)
-        servlet.emailService = Mock(EmailService)
+        service = new GroovyRunServiceImpl();
+
+        service.replicator = Mock(Replicator)
+        service.componentService = Mock(OsgiComponentService)
+        service.bundleContext = Mock(BundleContext)
+        service.configurationService = Mock(ConfigurationService)
+        service.queryBuilder = Mock(QueryBuilder)
+        service.emailService = Mock(EmailService)
+
+        servlet.groovyRunService = service;
 
         script = getScriptAsString("Script")
     }
